requirements_test.txt
pytest==7.1.2
pytest-django==4.5.2
responses==0.20.0
mypy==0.942
flake8==4.0.1
django-stubs==1.10.1
types-pysaml2==1.0.0
types-setuptools==57.4.14
types-pkg-resources==0.1.3
interrogate==1.5.0
coveralls==3.3.1
